requirements.txt
emails
psycopg
psycopg2-binary
emails
pydantic-settings
passlib
PyJWT
sqlmodel
sentry-sdk
fastapi[standard]
alembic